==> analysis-src/1_data_conversation.ipynb <==
#imports
import numpy as np
import pandas as pd
import random
import string

import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

import os
import csv
import datetime
from datetime import timedelta,datetime, date

# Перезапись CSV в Apache Parquet

transactions = pd.read_csv('transactions.csv')

transactions.START_DATE = pd.to_datetime(transactions.START_DATE)
transactions.END_DATE = pd.to_datetime(transactions.END_DATE)
transactions.DATE = pd.to_datetime(transactions.DATE)
transactions.EXP_TYPE = transactions.EXP_TYPE.astype("category")

transactions.TRANS_ID.duplicated().sum()

transactions.drop(columns= ["TRANS_ID"], inplace= True)

cl_to_ids = {}
for (i,v) in enumerate(transactions.CUST_ID.unique()):
  cl_to_ids[v] = i

transactions.CUST_ID = [cl_to_ids[x]  for x in transactions.CUST_ID]

print("Всего клиентов: {}".format(len(transactions['CUST_ID'].unique())))
print("Всего записей: {}".format(len(transactions)))

clients = pd.read_csv("customers.csv")
clients.head()

clients.PROD_TYPE = clients.PROD_TYPE.astype("category")
clients.CUST_ID = [cl_to_ids[x]  for x in clients.CUST_ID]

print("Всего клиентов: {}".format(len(clients['CUST_ID'].unique())))
print("Всего записей: {}".format(len(clients)))

#transactions.dropna(inplace= True)
#print("Всего клиентов: {}".format(len(transactions['CUST_ID'].unique())))
#print("Всего записей: {}".format(len(transactions)))

clients = clients.loc[clients.CUST_ID.isin(transactions.CUST_ID.unique()), :]
print("Всего клиентов: {}".format(len(clients['CUST_ID'].unique())))
print("Всего записей: {}".format(len(clients)))

transactions.YEAR = transactions.YEAR.round().fillna(-1).astype("int")
transactions.MONTH = transactions.MONTH.round().fillna(-1).astype("int")
transactions.DAY = transactions.DAY.round().fillna(-1).astype("int")

transactions.head()

clients.head()

transactions.EXP_TYPE.cat.categories[0] == 'Отели, мотели, курорты'

### латинские буквы в русских названиях!

import re
cats = transactions.EXP_TYPE.cat.categories.tolist()
[re.match("[, а-яА-Я]+", c) is not None   for c in cats]

cats[0] = 'Отели, мотели, курорты'
cats = [x.strip()   for x in cats]
transactions.EXP_TYPE.cat.categories = cats
transactions.EXP_TYPE = 
  transactions.EXP_TYPE.cat.reorder_categories(sorted(cats))

cats = clients.PROD_TYPE.cat.categories.tolist()
[re.match("[, а-яА-Я]+", c) is not None   for c in cats]

clients.PROD_TYPE.cat.categories = [x.strip()   for x in cats]

np.savetxt("exptypes.txt", transactions.EXP_TYPE.cat.categories.to_numpy(), 
           "%s")
np.savetxt("prodtypes.txt", clients.PROD_TYPE.cat.categories.to_numpy(), 
           "%s")

transactions.to_parquet("transactions.parquet")

clients.to_parquet("customers.parquet")
