==> analysis-src/4_colfilt_for_impfeedback.ipynb <==
#imports
import numpy as np
import pandas as pd
import random
import string

import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

import os
import csv
import datetime
from datetime import timedelta,datetime, date

from scipy import stats
from scipy.sparse import csr_matrix

import sklearn
from sklearn import metrics

import implicit
from implicit.nearest_neighbours import bm25_weight
from implicit.als import AlternatingLeastSquares

cu = pd.read_parquet("customers.parquet")
tr = pd.read_parquet("transactions_2_cleaned.parquet")

exptypes = tr.EXP_TYPE.unique()
prodtypes = cu.PROD_TYPE.unique()

cutoprod = {c:v.PROD_TYPE for (c,v) in cu.groupby("CUST_ID")}

grp = tr.loc[:, ["CUST_ID", "EXP_TYPE", "AMOUNT"]].groupby("CUST_ID")
l = []
lcu = []

for (k,v) in grp:
  df2 = v.groupby("EXP_TYPE").apply(lambda x: pd.Series([x.shape[0], 
    x.AMOUNT.sum()], index= ["n", "AMOUNT"]))
  df2.loc[:, "AMOUNT_prop"] = df2.AMOUNT / df2.AMOUNT.sum()
  df2.loc[:, "n_prop"] = df2.n / df2.n.sum()
  df2.loc[:, "AMOUNT_per_n"] = [0.0 if np.isnan(x) else x  
                                for x in df2.AMOUNT / df2.n]
  l.append(df2)
  lcu.append(k)

nc = len(exptypes)

datax = np.zeros((len(l), 3*len(exptypes)))
for (i,v) in enumerate(l):
  datax[i, range(nc)] = v.AMOUNT_prop
  datax[i, range(nc, nc + nc)] = v.n_prop
  datax[i, range(2*nc, 2*nc + nc)] = v.AMOUNT_per_n
  
datax_ap = datax[:, range(nc)]
datax_np = datax[:, range(nc, 2*nc)]
datax_apn = datax[:, range(2*nc, 3*nc)]

datay = np.array([ np.array([float((p == cutoprod[cu]).any())  for cu in lcu]) 
         for p in prodtypes]).transpose()

https://github.com/benfred/implicit/issues/125 

def prepare_model(datax):
  model = AlternatingLeastSquares(factors=34, regularization=0.05)
  datax2 = csr_matrix(datax)
  csrdatax = bm25_weight(datax2.T.tocsr(), K1=100, B=0.8).T.tocsr()
  model.fit(csrdatax, False)
  return(model)

def findsimilar_users(model, datax, newdata):
  newi = datax.shape[0]
  datax2 = csr_matrix(np.vstack([datax, newdata]))
  csrdatax = bm25_weight(datax2.T.tocsr(), K1=100, B=0.8).T.tocsr()
  model.partial_fit_users([newi], csr_matrix(datax2[[newi], :]))
  return(model.similar_users(newi, 10, [newi]))

#def average_products(datay, su):

model = prepare_model(datax)

su = findsimilar_users(model, datax, datax[12, :])
print(su)

datay[su[0],:]


def findsimilar_users2(datax, newdata):
  model = AlternatingLeastSquares(factors=34, regularization=0.05)
  newi = datax.shape[0]
  datax2 = csr_matrix(np.vstack([datax, newdata]))
  csrdatax = datax2#bm25_weight(datax2.T.tocsr(), K1=2, B=0.8).T.tocsr()
  model.fit(csrdatax, False)
  return(model.similar_users(newi, 10, [newi]))  

su = findsimilar_users2(datax, datax[[12], :])
print(su)

datay[su[0],:]


def findsimilar_users3(datax, newdata):
  model = AlternatingLeastSquares(factors=34, regularization=0.05)
  newi = datax.shape[0]
  datax2 = csr_matrix(np.vstack([datax, newdata]))
  csrdatax = bm25_weight(datax2, K1=2, B=0.8).tocsr()
  model.fit(csrdatax, False)
  return(model.similar_users(newi, 10, [newi]))  

su = findsimilar_users3(datax, datax[[12], :])
print(su)

datay[su[0],:]


su = findsimilar_users3(datax_ap, datax_ap[[12], :])
print(su)

datay[su[0],:]
