==> analysis-src/6_tests_for_streamlit.ipynb <==
#imports
import numpy as np
import pandas as pd
import random
import string

import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

import os
import csv
import datetime
from datetime import timedelta,datetime, date

from scipy import stats
import pickle

import catboost

import sklearn
from sklearn import metrics
from sklearn.linear_model import LogisticRegression

os.chdir("streamlit")

exptypes = [x.strip()   for x in open("exptypes.txt", "r").readlines()]
prodtypes = [x.strip()   for x in open("prodtypes.txt", "r").readlines()]

df = pd.read_csv("test.csv")

df.EXP_TYPE = pd.Categorical(df.EXP_TYPE, exptypes)

def transform_to_input_format(df):
  df2 = df.groupby("EXP_TYPE").apply(lambda x: pd.Series([x.shape[0], 
    x.AMOUNT.sum()], index= ["n", "AMOUNT"]))
  df2.loc[:, "AMOUNT_prop"] = df2.AMOUNT / df2.AMOUNT.sum()
  df2.loc[:, "n_prop"] = df2.n / df2.n.sum()
  df2.loc[:, "AMOUNT_per_n"] = [0.0 if np.isnan(x) else x  
                                for x in df2.AMOUNT / df2.n]
  nc = len(exptypes)
  datax = np.zeros((1, 3*len(exptypes)))
  datax[0, range(nc)] = df2.AMOUNT_prop
  datax[0, range(nc, nc + nc)] = df2.n_prop
  datax[0, range(2*nc, 2*nc + nc)] = df2.AMOUNT_per_n

  return(datax)

datax = transform_to_input_format(df)

def load_pretrained_models():
  def cbload(i):
    from_file = catboost.CatBoostClassifier()
    from_file.load_model("catboost_p{}.cbm".format(i))  
    return(from_file)

  cb_models = [cbload(i)   for i in range(10)]
  lr_models = pickle.load(open("logregr_models.pickle", 'rb'))
  cb_thrstats = pickle.load(open("catboost_thr.pickle", 'rb'))
  lr_thrstats = pickle.load(open("logregr_thr.pickle", 'rb'))
  return(cb_models, lr_models, cb_thrstats, lr_thrstats)

cb_models, lr_models, cb_thrstats, lr_thrstats = load_pretrained_models()

def predictions(datax):
  return(pd.DataFrame({
      "cb": [cb_models[i].predict_proba(datax)[0][1] 
             for i in range(len(prodtypes))],
      "lr": [lr_models[i].predict_proba(datax)[0][1] 
             for i in range(len(prodtypes))]
    }))
  
pr = predictions(datax)
pr

def find_approp_thresholds(thrstats, target, what): #what=1 -- prec, 2 -- recall
  r = []
  for i in range(len(thrstats[1])):
    l = thrstats[what][i] >= target
    if l.sum() > 0:
      j = np.where(l)[0][np.argmin(thrstats[what][i][l])]
    else:
      j = np.argmax(thrstats[what][i])
    r.append({"thr": thrstats[0][j], "prec": thrstats[1][i][j], 
              "recall": thrstats[2][i][j]})
  return pd.DataFrame(r)
      

find_approp_thresholds(cb_thrstats, 0.9, 1)

find_approp_thresholds(lr_thrstats, 0.9, 1)

targmetr = 1
targmetrval = 0.9
metrother = 2 if targmetr == 1 else 1

def find_approp_thresholds(thrstats, target, what): #what=1 -- prec, 2 -- recall
  r = []
  for i in range(len(thrstats[1])):
    l = thrstats[what][i] >= target
    if l.sum() > 0:
      j = np.where(l)[0][np.argmin(thrstats[what][i][l])]
    else:
      j = np.argmax(thrstats[what][i])
    r.append({"thr": thrstats[0][j], "prec": thrstats[1][i][j], 
              "recall": thrstats[2][i][j]})
  return pd.DataFrame(r)
cbt = find_approp_thresholds(cb_thrstats, targmetrval, targmetr)
lrt = find_approp_thresholds(lr_thrstats, targmetrval, targmetr)

print(cbt)
print(lrt)

r = []
for i in range(pr.shape[0]):
  if cbt.iloc[i, metrother] > lrt.iloc[i, metrother]:
    if pr.cb[i] >= cbt.thr[i]:
      r.append(prodtypes[i])
  else:
    if pr.lr[i] >= lrt.thr[i]:
      r.append(prodtypes[i])
print(r)

pd.DataFrame({
    "product": prodtypes,   "score": pr.cb,   "score_threshold": cbt.thr,   
    "rule precision": cbt.prec, "rule recall": cbt.recall,
  })

pd.DataFrame({
    "product": prodtypes,   "score": pr.lr,   "score_threshold": lrt.thr,   
    "rule precision": lrt.prec, "rule recall": lrt.recall,
  })
